Successfully got in contact with another former student, who gave permission for their game to be contributed to the Community repo.
Presenting Dawson Morey's "HauntyGhost", a "flappy bird"-style game written entirely in Vircon32 assembly (v24.2.4 or later to build), written in Spring 2024.

This is among the simplest to operate: using only the UP, LEFT, and RIGHT direction buttons on the controller, which makes it great for general audience outreach
events.