Update bios without logo debug
* Added Readme.md, parameter tweak to single parameter instructions, updated Readme.txt

* Formatting tweaks to Readme.md

* Removing plaintext `Readme.txt`.

Updating markdown `Readme.md` to reflect current changes and other useful BIOS information.

Updating screenshot to include portname translation.

BiosWithoutLogoDebug.c now has portname translation support for both `IN` and `OUT`
instructions; this covers all ports from all seven categories.

NOTE that the port COMMANDS are still not translated- you will see a raw hex value.

* Initial attempt at bounds checking: first tests did not work, will need to investigate.

The infrastructure is in place, and I hope ultimately usable.

* Accidentally lost all PortName translation logic. Reimplemented.

* Discovered another bug: MOV addressing modes 5-7 were not being directed properly
in program logic.

The flag variable was used to set a special value of 2, which when applied against
the check for immediate value, will kick us to the else in order to display the register.

* Checking lower bound is still broken:

* will work if absolute first instruction (that has error)
* will not work if there is one instruction prior to errored instruction
* works if there are two instructions prior to the errored instruction (2+ average case)

This may be the last bug holding things up.

* Reviewing code, polishing formatting for function-level consistency

* possibly have proper lower bounds checking fully working

* updating Readme.md to correct To-Do section, now that bounds checking appears to work